Move getSwatchImageUrl into separate file
